Fix type casting for ModuleInformationLength comparison

+++ ntoskrnl/ex/sysinfo.c
@@ -6034,7 +6034,7 @@ ExpQueryModuleInformation (
     if (ARGUMENT_PRESENT(ReturnLength)) {
         *ReturnLength = RequiredLength;
     }
-    if (ModuleInformationLength >= FIELD_OFFSET( RTL_PROCESS_MODULES, Modules )) {
+    if (ModuleInformationLength >= (ULONG)FIELD_OFFSET(RTL_PROCESS_MODULES, Modules)) {
         ModuleInformation->NumberOfModules = NumberOfModules;
     } else {
         Status = STATUS_INFO_LENGTH_MISMATCH;
